+++ age/src/protocol.rs
@@ -7,9 +7,10 @@ use std::iter;
 
 use crate::{
     error::Error,
-    format::{oil_the_joint, scrypt, Header, HeaderV1, RecipientStanza},
-    keys::{FileKey, RecipientKey},
+    format::{oil_the_joint, Header, HeaderV1, RecipientStanza},
+    keys::FileKey,
     primitives::stream::{PayloadKey, Stream, StreamWriter},
+    scrypt, Recipient,
 };
 
 #[cfg(feature = "async")]
@@ -55,7 +56,7 @@ pub trait Callbacks {
 /// Handles the various types of age encryption.
 enum EncryptorType {
     /// Encryption to a list of recipients identified by keys.
-    Keys(Vec<RecipientKey>),
+    Keys(Vec<Box<dyn Recipient>>),
     /// Encryption to a passphrase.
     Passphrase(SecretString),
 }
@@ -66,17 +67,17 @@ pub struct Encryptor(EncryptorType);
 impl Encryptor {
     /// Returns an `Encryptor` that will create an age file encrypted to a list of
     /// recipients.
-    pub fn with_recipients(recipients: Vec<RecipientKey>) -> Self {
+    pub fn with_recipients(recipients: Vec<Box<dyn Recipient>>) -> Self {
         Encryptor(EncryptorType::Keys(recipients))
     }
 
     /// Returns an `Encryptor` that will create an age file encrypted with a passphrase.
     ///
     /// This API should only be used with a passphrase that was provided by (or generated
-    /// for) a human. For programmatic use cases, instead generate a [`SecretKey`] and
-    /// then use [`Encryptor::with_recipients`].
+    /// for) a human. For programmatic use cases, instead generate an [`x25519::Identity`]
+    /// and then use [`Encryptor::with_recipients`].
     ///
-    /// [`SecretKey`]: crate::keys::SecretKey
+    /// [`x25519::Identity`]: crate::x25519::Identity
     pub fn with_user_passphrase(passphrase: SecretString) -> Self {
         Encryptor(EncryptorType::Passphrase(passphrase))
     }
@@ -222,7 +223,7 @@ mod tests {
     use std::iter;
 
     use super::{Decryptor, Encryptor};
-    use crate::{identity::IdentityFile, keys::RecipientKey, Identity};
+    use crate::{identity::IdentityFile, ssh, x25519, Identity, Recipient};
 
     #[cfg(feature = "async")]
     use futures::{
@@ -235,7 +236,7 @@ mod tests {
     use futures_test::task::noop_context;
 
     fn recipient_round_trip(
-        recipients: Vec<RecipientKey>,
+        recipients: Vec<Box<dyn Recipient>>,
         identities: impl Iterator<Item = Box<dyn Identity>>,
     ) {
         let test_msg = b"This is a test message. For testing.";
@@ -261,7 +262,7 @@ mod tests {
 
     #[cfg(feature = "async")]
     fn recipient_async_round_trip(
-        recipients: Vec<RecipientKey>,
+        recipients: Vec<Box<dyn Recipient>>,
         identities: impl Iterator<Item = Box<dyn Identity>>,
     ) {
         let test_msg = b"This is a test message. For testing.";
@@ -339,19 +340,25 @@ mod tests {
 
     #[test]
     fn x25519_round_trip() {
-        let buf = BufReader::new(crate::keys::tests::TEST_SK.as_bytes());
+        let buf = BufReader::new(crate::x25519::tests::TEST_SK.as_bytes());
         let sk = IdentityFile::from_buffer(buf).unwrap();
-        let pk: RecipientKey = crate::keys::tests::TEST_PK.parse().unwrap();
-        recipient_round_trip(vec![pk], iter::once(Box::new(sk) as Box<dyn Identity>));
+        let pk: x25519::Recipient = crate::x25519::tests::TEST_PK.parse().unwrap();
+        recipient_round_trip(
+            vec![Box::new(pk)],
+            iter::once(Box::new(sk) as Box<dyn Identity>),
+        );
     }
 
     #[cfg(feature = "async")]
     #[test]
     fn x25519_async_round_trip() {
-        let buf = BufReader::new(crate::keys::tests::TEST_SK.as_bytes());
+        let buf = BufReader::new(crate::x25519::tests::TEST_SK.as_bytes());
         let sk = IdentityFile::from_buffer(buf).unwrap();
-        let pk: RecipientKey = crate::keys::tests::TEST_PK.parse().unwrap();
-        recipient_async_round_trip(vec![pk], iter::once(Box::new(sk) as Box<dyn Identity>));
+        let pk: x25519::Recipient = crate::x25519::tests::TEST_PK.parse().unwrap();
+        recipient_async_round_trip(
+            vec![Box::new(pk)],
+            iter::once(Box::new(sk) as Box<dyn Identity>),
+        );
     }
 
     #[test]
@@ -383,10 +390,13 @@ mod tests {
     fn ssh_rsa_round_trip() {
         let buf = BufReader::new(crate::ssh::identity::tests::TEST_SSH_RSA_SK.as_bytes());
         let sk = crate::ssh::identity::Identity::from_buffer(buf, None).unwrap();
-        let pk: RecipientKey = crate::ssh::recipient::tests::TEST_SSH_RSA_PK
+        let pk: ssh::Recipient = crate::ssh::recipient::tests::TEST_SSH_RSA_PK
             .parse()
             .unwrap();
-        recipient_round_trip(vec![pk], iter::once(Box::new(sk) as Box<dyn Identity>));
+        recipient_round_trip(
+            vec![Box::new(pk)],
+            iter::once(Box::new(sk) as Box<dyn Identity>),
+        );
     }
 
     #[cfg(feature = "async")]
@@ -394,20 +404,26 @@ mod tests {
     fn ssh_rsa_async_round_trip() {
         let buf = BufReader::new(crate::ssh::identity::tests::TEST_SSH_RSA_SK.as_bytes());
         let sk = crate::ssh::identity::Identity::from_buffer(buf, None).unwrap();
-        let pk: RecipientKey = crate::ssh::recipient::tests::TEST_SSH_RSA_PK
+        let pk: ssh::Recipient = crate::ssh::recipient::tests::TEST_SSH_RSA_PK
             .parse()
             .unwrap();
-        recipient_async_round_trip(vec![pk], iter::once(Box::new(sk) as Box<dyn Identity>));
+        recipient_async_round_trip(
+            vec![Box::new(pk)],
+            iter::once(Box::new(sk) as Box<dyn Identity>),
+        );
     }
 
     #[test]
     fn ssh_ed25519_round_trip() {
         let buf = BufReader::new(crate::ssh::identity::tests::TEST_SSH_ED25519_SK.as_bytes());
         let sk = crate::ssh::identity::Identity::from_buffer(buf, None).unwrap();
-        let pk: RecipientKey = crate::ssh::recipient::tests::TEST_SSH_ED25519_PK
+        let pk: ssh::Recipient = crate::ssh::recipient::tests::TEST_SSH_ED25519_PK
             .parse()
             .unwrap();
-        recipient_round_trip(vec![pk], iter::once(Box::new(sk) as Box<dyn Identity>));
+        recipient_round_trip(
+            vec![Box::new(pk)],
+            iter::once(Box::new(sk) as Box<dyn Identity>),
+        );
     }
 
     #[cfg(feature = "async")]
@@ -415,9 +431,12 @@ mod tests {
     fn ssh_ed25519_async_round_trip() {
         let buf = BufReader::new(crate::ssh::identity::tests::TEST_SSH_ED25519_SK.as_bytes());
         let sk = crate::ssh::identity::Identity::from_buffer(buf, None).unwrap();
-        let pk: RecipientKey = crate::ssh::recipient::tests::TEST_SSH_ED25519_PK
+        let pk: ssh::Recipient = crate::ssh::recipient::tests::TEST_SSH_ED25519_PK
             .parse()
             .unwrap();
-        recipient_async_round_trip(vec![pk], iter::once(Box::new(sk) as Box<dyn Identity>));
+        recipient_async_round_trip(
+            vec![Box::new(pk)],
+            iter::once(Box::new(sk) as Box<dyn Identity>),
+        );
     }
 }